crates/pygreeter/pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "pygreeter"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
    "asyncio~=3.4.3",
    "opentelemetry-api~=1.28.0",
    "opentelemetry-exporter-otlp~=1.28.0",
]
[tool.maturin]
features = ["pyo3/extension-module"]
